Make CreateTableOpt public to ease creation of custom CatalogIO implementations
